requirements.txt
numpy>=1.21.0
pycryptodome>=3.15.0
scapy>=2.4.5